+++ src/cli/commands/private-cmds/_graph-cmd.ts
@@ -4,7 +4,9 @@ import { graph } from '../../../api/scope';
 import { migrate } from '../../../api/consumer';
 import logger from '../../../logger/logger';
 import { checkVersionCompatibilityOnTheServer } from '../../../scope/network/check-version-compatibility';
+import clientSupportCompressedCommand from '../../../utils/ssh/client-support-compressed-command';
 
+let compressResponse;
 // eslint-disable-next-line @typescript-eslint/class-name-casing
 export default class _Graph extends Command {
   name = '_graph <path> <args>';
@@ -15,6 +17,7 @@ export default class _Graph extends Command {
 
   action([path, args]: [string, string]): Promise<any> {
     const { payload, headers } = unpackCommand(args);
+    compressResponse = clientSupportCompressedCommand(headers.version);
     checkVersionCompatibilityOnTheServer(headers.version);
     logger.info('Checking if a migration is needed');
     const scopePath = fromBase64(path);
@@ -24,6 +27,6 @@ export default class _Graph extends Command {
   }
 
   report(str: string): string {
-    return packCommand(buildCommandMessage(str));
+    return packCommand(buildCommandMessage(str, undefined, compressResponse), true, compressResponse);
   }
 }